[REFACTOR] conver 수정 및 스웨거 수정
